requirements.txt
torch==2.0.0
pandas==1.5.3
numpy==1.24.2
Pillow==9.4.0
scipy==1.10.1
tensorboard==2.12.2
torchvision==0.15.1
tqdm==4.64.1
wilds==2.0.0
matplotlib==3.7.1
